Added utils module for stereochemistry
